+++ src/job.ts
@@ -39,6 +39,7 @@ import {
 } from './bigquery';
 import {RowMetadata} from './table';
 import bigquery from './types';
+import {logger} from './logger';
 
 export type JobMetadata = bigquery.IJob;
 export type JobOptions = JobRequest<JobMetadata>;
@@ -50,7 +51,12 @@ export type QueryResultsOptions = {
   job?: Job;
   wrapIntegers?: boolean | IntegerTypeCastOptions;
   parseJSON?: boolean;
-} & PagedRequest<bigquery.jobs.IGetQueryResultsParams>;
+} & PagedRequest<bigquery.jobs.IGetQueryResultsParams> & {
+    /**
+     * internal properties
+     */
+    _cachedRows?: any[];
+  };
 
 /**
  * @callback QueryResultsCallback
@@ -379,6 +385,11 @@ class Job extends Operation {
     );
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  private trace_(msg: string, ...otherArgs: any[]) {
+    logger(`[job][${this.id}]`, msg, ...otherArgs);
+  }
+
   /**
    * @callback CancelCallback
    * @param {?Error} err Request error, if any.
@@ -529,12 +540,18 @@ class Job extends Operation {
       typeof optionsOrCallback === 'object' ? optionsOrCallback : {};
     const callback =
       typeof optionsOrCallback === 'function' ? optionsOrCallback : cb;
-
     const qs = extend(
       {
         location: this.location,
+        'formatOptions.useInt64Timestamp': true,
       },
       options
+    );
+    this.trace_(
+      '[getQueryResults]',
+      this.id,
+      options.pageToken,
+      options.startIndex
     );
 
     const wrapIntegers = qs.wrapIntegers ? qs.wrapIntegers : false;
@@ -546,6 +563,18 @@ class Job extends Operation {
 
     const timeoutOverride =
       typeof qs.timeoutMs === 'number' ? qs.timeoutMs : false;
+
+    if (options._cachedRows) {
+      let nextQuery: QueryResultsOptions | null = null;
+      if (options.pageToken) {
+        nextQuery = Object.assign({}, options, {
+          pageToken: options.pageToken,
+        });
+        delete nextQuery._cachedRows;
+      }
+      callback!(null, options._cachedRows, nextQuery);
+      return;
+    }
 
     this.bigQuery.request(
       {
@@ -582,6 +611,7 @@ class Job extends Operation {
             return;
           }
         } else if (resp.pageToken) {
+          this.trace_('[getQueryResults] has more pages', resp.pageToken);
           // More results exist.
           nextQuery = Object.assign({}, options, {
             pageToken: resp.pageToken,