+++ test/table.ts
@@ -173,9 +173,9 @@ describe('BigQuery/Table', () => {
 
   const DATASET = {
     id: 'dataset-id',
+    projectId: 'project-id',
     createTable: util.noop,
     bigQuery: {
-      projectId: 'project-id',
       job: (id: string) => {
         return {id};
       },
@@ -721,12 +721,12 @@ describe('BigQuery/Table', () => {
               c: 'd',
               destinationTable: {
                 datasetId: DEST_TABLE.dataset.id,
-                projectId: DEST_TABLE.bigQuery.projectId,
+                projectId: DEST_TABLE.dataset.projectId,
                 tableId: DEST_TABLE.id,
               },
               sourceTable: {
                 datasetId: table.dataset.id,
-                projectId: table.bigQuery.projectId,
+                projectId: table.dataset.projectId,
                 tableId: table.id,
               },
             },
@@ -842,13 +842,13 @@ describe('BigQuery/Table', () => {
               c: 'd',
               destinationTable: {
                 datasetId: table.dataset.id,
-                projectId: table.bigQuery.projectId,
+                projectId: table.dataset.projectId,
                 tableId: table.id,
               },
               sourceTables: [
                 {
                   datasetId: SOURCE_TABLE.dataset.id,
-                  projectId: SOURCE_TABLE.bigQuery.projectId,
+                  projectId: SOURCE_TABLE.dataset.projectId,
                   tableId: SOURCE_TABLE.id,
                 },
               ],
@@ -867,12 +867,12 @@ describe('BigQuery/Table', () => {
         assert.deepStrictEqual(reqOpts.configuration!.copy!.sourceTables, [
           {
             datasetId: SOURCE_TABLE.dataset.id,
-            projectId: SOURCE_TABLE.bigQuery.projectId,
+            projectId: SOURCE_TABLE.dataset.projectId,
             tableId: SOURCE_TABLE.id,
           },
           {
             datasetId: SOURCE_TABLE.dataset.id,
-            projectId: SOURCE_TABLE.bigQuery.projectId,
+            projectId: SOURCE_TABLE.dataset.projectId,
             tableId: SOURCE_TABLE.id,
           },
         ]);
@@ -1002,7 +1002,7 @@ describe('BigQuery/Table', () => {
       table.bigQuery.createJob = (reqOpts: JobOptions) => {
         assert.deepStrictEqual(reqOpts.configuration!.extract!.sourceTable, {
           datasetId: table.dataset.id,
-          projectId: table.bigQuery.projectId,
+          projectId: table.dataset.projectId,
           tableId: table.id,
         });
 
@@ -1685,14 +1685,14 @@ describe('BigQuery/Table', () => {
                 a: 'b',
                 c: 'd',
                 destinationTable: {
-                  projectId: table.bigQuery.projectId,
+                  projectId: table.dataset.projectId,
                   datasetId: table.dataset.id,
                   tableId: table.id,
                 },
               },
             },
             jobReference: {
-              projectId: table.bigQuery.projectId,
+              projectId: table.dataset.projectId,
               jobId: fakeJobId,
               location: undefined,
             },
@@ -1711,7 +1711,7 @@ describe('BigQuery/Table', () => {
           const uri =
             table.bigQuery.apiEndpoint +
             '/upload/bigquery/v2/projects/' +
-            table.bigQuery.projectId +
+            table.dataset.projectId +
             '/jobs';
           assert.strictEqual(options.request.uri, uri);
           done();
@@ -1783,6 +1783,7 @@ describe('BigQuery/Table', () => {
           jobReference: {
             jobId: 'job-id',
             location: 'location',
+            projectId: 'project-id',
           },
           a: 'b',
           c: 'd',
@@ -1792,6 +1793,7 @@ describe('BigQuery/Table', () => {
           assert.strictEqual(id, metadata.jobReference!.jobId);
           assert.deepStrictEqual(options, {
             location: metadata.jobReference!.location,
+            projectId: metadata.jobReference!.projectId,
           });
           return fakeJob;
         };
@@ -1979,7 +1981,10 @@ describe('BigQuery/Table', () => {
 
       table.request = (reqOpts: DecorateRequestOptions, callback: Function) => {
         assert.strictEqual(reqOpts.uri, '/data');
-        assert.strictEqual(reqOpts.qs, options);
+        assert.deepStrictEqual(reqOpts.qs, {
+          ...options,
+          'formatOptions.useInt64Timestamp': true,
+        });
         callback(null, {});
       };
 
@@ -2129,7 +2134,10 @@ describe('BigQuery/Table', () => {
       table.metadata = {schema: {}};
 
       table.request = (reqOpts: DecorateRequestOptions, callback: Function) => {
-        callback(null, {pageToken});
+        callback(null, {
+          'formatOptions.useInt64Timestamp': true,
+          pageToken,
+        });
       };
 
       table.getRows(options, (err: Error, rows: {}, nextQuery: {}) => {
@@ -2137,6 +2145,7 @@ describe('BigQuery/Table', () => {
         assert.deepStrictEqual(nextQuery, {
           a: 'b',
           c: 'd',
+          'formatOptions.useInt64Timestamp': true,
           pageToken,
         });
         // Original object isn't affected.
@@ -2257,7 +2266,9 @@ describe('BigQuery/Table', () => {
       const merged = [{name: 'stephen'}];
 
       table.request = (reqOpts: DecorateRequestOptions, callback: Function) => {
-        assert.deepStrictEqual(reqOpts.qs, {});
+        assert.deepStrictEqual(reqOpts.qs, {
+          'formatOptions.useInt64Timestamp': true,
+        });
         callback(null, {});
       };
 
@@ -2279,7 +2290,9 @@ describe('BigQuery/Table', () => {
       const merged = [{name: 'stephen'}];
 
       table.request = (reqOpts: DecorateRequestOptions, callback: Function) => {
-        assert.deepStrictEqual(reqOpts.qs, {});
+        assert.deepStrictEqual(reqOpts.qs, {
+          'formatOptions.useInt64Timestamp': true,
+        });
         callback(null, {});
       };